Tambah UI Product + Detail Pembayaran

+++ app/src/main/java/com/bca/bsi/utils/Utils.java
@@ -45,7 +45,7 @@ public class Utils {
         }
     }
 
-    public static String priceFormat(long totalPrice) {
+    public static String priceFormat(double totalPrice) {
         DecimalFormat formatter = new DecimalFormat("#,###,###");
         return formatter.format(totalPrice);
     }
